Code refactor for the application.

+++ src/main/java/pl/wasik/damian/spring/app/fleet/management/system/fleet/repositories/VehicleMakeRepository.java
@@ -1,9 +1,11 @@
 package pl.wasik.damian.spring.app.fleet.management.system.fleet.repositories;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.wasik.damian.spring.app.fleet.management.system.fleet.models.VehicleMake;
 
 @Repository
-public interface VehicleMakeRepository extends JpaRepository<VehicleMake, Long> {
+public interface VehicleMakeRepository extends JpaRepository<VehicleMake, Integer> {
+
 }
