+++ src/main/java/pl/wasik/damian/spring/app/fleet/management/system/fleet/repositories/VehicleStatusRepository.java
@@ -1,10 +1,10 @@
 package pl.wasik.damian.spring.app.fleet.management.system.fleet.repositories;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pl.wasik.damian.spring.app.fleet.management.system.fleet.models.VehicleStatus;
 
 @Repository
-public interface VehicleStatusRepository extends JpaRepository<VehicleStatus, Long> {
+public interface VehicleStatusRepository extends JpaRepository<VehicleStatus, Integer> {
+
 }
